+++ tests/resource/setup.py
@@ -8,7 +8,8 @@ from typing import Any, Type
 from fastapi import FastAPI
 
 from apexdevkit.fastapi import FastApiBuilder
-from apexdevkit.fastapi.router import RestfulRouter, RestfulServiceBuilder
+from apexdevkit.fastapi.builder import RestfulServiceBuilder
+from apexdevkit.fastapi.router import RestfulRouter
 from apexdevkit.testing import RestfulName
 from apexdevkit.testing.fake import FakeResource
 from tests.resource.sample_api import (
